Refactor: 게시글 페이지 폴더 구조 수정 및 리팩토링
