+++ src/pages/editor/extensions.ts
@@ -1,0 +1,79 @@
+import Title from '@/components/editor/Title';
+import Color from '@tiptap/extension-color';
+import Document from '@tiptap/extension-document';
+import Heading from '@tiptap/extension-heading';
+import Image from '@tiptap/extension-image';
+import ListItem from '@tiptap/extension-list-item';
+import Placeholder from '@tiptap/extension-placeholder';
+import TextStyle from '@tiptap/extension-text-style';
+import Underline from '@tiptap/extension-underline';
+import { Extensions, textblockTypeInputRule } from '@tiptap/react';
+import StarterKit from '@tiptap/starter-kit';
+
+const DocumentWithTitle = Document.extend({
+  content: 'title block+',
+});
+
+const adjustLevel = (level: any) => (level == 1 ? 2 : level);
+const CustomHeading = Heading.extend({
+  parseHTML() {
+    return this.options.levels.map((level) => ({
+      tag: `h${level}`,
+      attrs: { level: adjustLevel(level) },
+    }));
+  },
+  addKeyboardShortcuts() {
+    return this.options.levels.reduce(
+      (items, level) => ({
+        ...items,
+        ...{
+          [`Mod-Alt-${level}`]: () =>
+            this.editor.commands.toggleHeading({
+              level: adjustLevel(level),
+            }),
+        },
+      }),
+      {}
+    );
+  },
+  addInputRules() {
+    return this.options.levels.map((level) => {
+      return textblockTypeInputRule({
+        find: new RegExp(`^(#{1,${level}})\\s$`),
+        type: this.type,
+        getAttributes: {
+          level: adjustLevel(level),
+        },
+      });
+    });
+  },
+});
+
+export const extensions: Extensions = [
+  Color.configure({ types: [TextStyle.name, ListItem.name] }),
+  TextStyle.configure({ types: [ListItem.name] } as any),
+  StarterKit.configure({
+    orderedList: {
+      keepMarks: true,
+      keepAttributes: false,
+    },
+    bulletList: {
+      keepMarks: true,
+      keepAttributes: false,
+    },
+    document: false,
+    heading: false,
+  }),
+  Placeholder.configure({
+    showOnlyCurrent: false,
+    placeholder: ({ node }) => {
+      if (node.type.name === 'title') return '제목을 입력하세요';
+      return '내용을 입력하세요';
+    },
+  }),
+  DocumentWithTitle,
+  Title,
+  CustomHeading,
+  Underline,
+  Image,
+];
