+++ src/pages/editor/index.tsx
@@ -1,96 +1,19 @@
 import './styles.scss';
 
-import { EditorContent, useEditor, textblockTypeInputRule } from '@tiptap/react';
-import { Color } from '@tiptap/extension-color';
-import ListItem from '@tiptap/extension-list-item';
-import TextStyle from '@tiptap/extension-text-style';
-import Underline from '@tiptap/extension-underline';
-import Document from '@tiptap/extension-document';
-import Heading from '@tiptap/extension-heading';
-import Placeholder from '@tiptap/extension-placeholder';
-import Image from '@tiptap/extension-image';
-import StarterKit from '@tiptap/starter-kit';
+import { EditorContent, useEditor } from '@tiptap/react';
 
-import MenuBar from './MenuBar';
-import Title from './Title';
+import MenuBar from './components/MenuBar';
 
-import useBlobUrl from '../../hooks/useBlobUrl';
-import { createBlob, uploadImageToFirebase } from '../../hooks/useImageConvert';
-import Button from '@/components/atoms/button';
 import { instance } from '@/api/client';
-import { useRecoilValue } from 'recoil';
+import Button from '@/components/atoms/button';
 import { userState } from '@/store/store';
+import { AxiosError } from 'axios';
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { AxiosError } from 'axios';
-
-const DocumentWithTitle = Document.extend({
-  content: 'title block+',
-});
-
-const adjustLevel = (level: any) => (level == 1 ? 2 : level);
-const CustomHeading = Heading.extend({
-  parseHTML() {
-    return this.options.levels.map((level) => ({
-      tag: `h${level}`,
-      attrs: { level: adjustLevel(level) },
-    }));
-  },
-  addKeyboardShortcuts() {
-    return this.options.levels.reduce(
-      (items, level) => ({
-        ...items,
-        ...{
-          [`Mod-Alt-${level}`]: () =>
-            this.editor.commands.toggleHeading({
-              level: adjustLevel(level),
-            }),
-        },
-      }),
-      {}
-    );
-  },
-  addInputRules() {
-    return this.options.levels.map((level) => {
-      return textblockTypeInputRule({
-        find: new RegExp(`^(#{1,${level}})\\s$`),
-        type: this.type,
-        getAttributes: {
-          level: adjustLevel(level),
-        },
-      });
-    });
-  },
-});
-
-const extensions = [
-  Color.configure({ types: [TextStyle.name, ListItem.name] }),
-  TextStyle.configure({ types: [ListItem.name] } as any),
-  StarterKit.configure({
-    orderedList: {
-      keepMarks: true,
-      keepAttributes: false,
-    },
-    bulletList: {
-      keepMarks: true,
-      keepAttributes: false,
-    },
-    document: false,
-    heading: false,
-  }),
-  Placeholder.configure({
-    showOnlyCurrent: false,
-    placeholder: ({ node }) => {
-      if (node.type.name === 'title') return '제목을 입력하세요';
-      return '내용을 입력하세요';
-    },
-  }),
-  DocumentWithTitle,
-  Title,
-  CustomHeading,
-  Underline,
-  Image,
-];
+import { useRecoilValue } from 'recoil';
+import useBlobUrl from '../../hooks/useBlobUrl';
+import { createBlob, uploadImageToFirebase } from '../../hooks/useImageConvert';
+import { extensions } from './extensions';
 
 interface Props {
   oldContent?: string;
@@ -227,10 +150,10 @@ const TextEditor = ({ oldContent, oldTitle, mode, postId }: Props) => {
           }}
         />
         <label htmlFor="anonymity">익명</label>
-        <Button type="button" className="conpact-red-200" show onClick={handleCancel}>
+        <Button type="button" className="compact-red-200" show onClick={handleCancel}>
           취소
         </Button>
-        <Button type="button" className="conpact-red-200" show onClick={handlePublish}>
+        <Button type="button" className="compact-red-200" show onClick={handlePublish}>
           {mode === 'edit' ? '수정' : '발행'}
         </Button>
       </div>
