+++ src/pages/editor/styles.scss
